requirements.txt
Appium-Python-Client==2.7.1
async-generator==1.10
attrs==22.1.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
distlib==0.3.6
filelock==3.8.0
Flask==2.2.2
h11==0.14.0
idna==3.4
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
outcome==1.2.0
packaging==21.3
platformdirs==2.5.4
pluggy==1.0.0
pycparser==2.21
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.2.0
pytest-env==0.8.1
pytest-runner==6.0.0
pytest-variables==2.0.0
requests==2.28.1
selenium==4.6.1
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.12
virtualenv==20.17.0
Werkzeug==2.2.2
wsproto==1.2.0
